+++ lib/src/app/adapters/adapter_stub.dart
@@ -1,0 +1,3 @@
+import '../app_settings.dart';
+
+void postConfigAppSettings(AppSettings settings) {}