+++ lib/src/app/http_client/adapters/browser_adapter.dart
@@ -1,12 +1,12 @@
-import 'package:dio/adapter_browser.dart';
-import 'package:dio/dio.dart';
+// import 'package:dio/adapter_browser.dart';
+// import 'package:dio/dio.dart';
 
-extension AppHttpDioAdapter on Dio {
-  void allowSelfSignedCert() {}
-  void allowWithCredential() {
-    (httpClientAdapter as BrowserHttpClientAdapter).withCredentials = true;
-  }
+// extension AppHttpDioAdapter on Dio {
+//   void allowSelfSignedCert() {}
+//   void allowWithCredential() {
+//     (httpClientAdapter as BrowserHttpClientAdapter).withCredentials = true;
+//   }
 
-  void enableCookieManager() {}
-  void clearCookies() {}
-}
+//   void enableCookieManager() {}
+//   void clearCookies() {}
+// }