+++ lib/src/app/http_client/adapters/http_adapter_stub.dart
@@ -1,0 +1,8 @@
+import 'package:dio/dio.dart';
+
+extension AppHttpDioAdapter on Dio {
+  void allowSelfSignedCert() {}
+  void allowWithCredential() {}
+  void enableCookieManager() {}
+  void clearCookies() {}
+}