+++ lib/src/app/http_client/adapters/stub_adapter.dart
@@ -1,8 +1,8 @@
-import 'package:dio/dio.dart';
+// import 'package:dio/dio.dart';
 
-extension AppHttpDioAdapter on Dio {
-  void allowSelfSignedCert() {}
-  void allowWithCredential() {}
-  void enableCookieManager() {}
-  void clearCookies() {}
-}
+// extension AppHttpDioAdapter on Dio {
+//   void allowSelfSignedCert() {}
+//   void allowWithCredential() {}
+//   void enableCookieManager() {}
+//   void clearCookies() {}
+// }