+++ test/main_test.dart
@@ -133,6 +133,7 @@ void main() {
         'jasonlaw',
         'Jason Law',
         'jason.cclaw@gmail.com',
+        tenantId: "XYZ",
         roles: 'SystemAdmin,ABC, CDE  ,ABC, ',
       );