+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.6.6
+* AppSettings baseUrl will follow the hosting environment in production
+* Added extension gqlFr to simplify defination of fragment
+* Move CredentialActions to CredentialService
+
 ## 0.6.5
 * Fixed XFileImage deprecated issue